requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
blobfile==3.0.0
brotlicffi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_b5py2dr5ag/croot/brotlicffi_1736182482518/work
cachetools==5.5.2
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_barck9k8fe/croot/certifi_1754570651396/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e4xd9yd9i2/croot/cffi_1736182819442/work
chardet==3.0.4
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
databricks-sdk==0.64.0
dataclasses-json==0.6.7
datasets==4.0.0
datasketch==1.6.5
Deprecated==1.2.18
dill==0.3.8
docker==7.1.0
effdet==0.4.1
emoji==2.15.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
FlashRank==0.2.10
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-translate==3.21.1
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
h2==3.2.0
hf-xet==1.1.7
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.7
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-milvus==0.1.6
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.31
lm-dataformat==0.0.20
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
ml_dtypes==0.5.3
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.23.0
opencv-python==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==24.2
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
peft==0.17.1
pi_heif==1.1.0
pikepdf==9.11.0
pillow==11.3.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
pymilvus==2.6.2
pyparsing==3.2.3
pypdf==6.1.1
pypdfium2==4.30.0
PySocks @ file:///Users/cbousseau/work/recipes/ci_py311/pysocks_1677906386870/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_87pebw_aae/croot/regex_1736541331314/work
requests @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_6fxef5l5oe/croot/requests_1750426350797/work
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.1
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.35
sqlparse==0.5.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_ec524s5k93/croot/tiktoken_1748479166250/work
timm==1.0.20
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.0
trl==0.22.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3 @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_95v32ztwo_/croot/urllib3_1750775480393/work
uvicorn==0.35.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
fastapi
uvicorn[standard]
httpx
elasticsearch>=8.0.0
psycopg2-binary
SQLAlchemy
alembic
python-dotenv
redis
rq
PyJWT
pydantic
python-multipart
requests
